requirements.txt
colorama==0.4.4
pysmb==1.2.8
